[#17] test: 스키마에 맞지 않은 필드 타입 변경
Long -> Integer

+++ src/test/java/dabang/star/cafe/api/ManagerApiTest.java
@@ -30,7 +30,7 @@ class ManagerApiTest {
     private final String MANAGER_NAME = "name";
     private final String MANAGER_PASSWORD = "password";
     private final Long MANAGER_ID = 1L;
-    private final Long OFFICE_ID = 1L;
+    private final Integer OFFICE_ID = 1;
 
     @Autowired
     private MockMvc mockMvc;
